Refactor out Windows specific includes into Platform.h, refactor out IWinToastHandler into its own header
Refactor out Windows specific includes into Platform.h, refactor out IWinToastHandler into its own header
